# openapi.yaml
openapi: 3.0.3
info:
  title: Oh Your Ear API
  description: API for the Oh Your Ear ear training application
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local development
paths:
...
  /me/plan:
    get:
      operationId: getStudyPlan
      summary: Get the study plan and today's progress for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Study plan with today's progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updateStudyPlan
      summary: Create or replace the study plan for the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyPlanUpdate'
      responses:
        '200':
          description: Saved study plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/practice-records:
    post:
      operationId: createPracticeRecord
      summary: Record the result of one answered question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeRecordRequest'
      responses:
        '201':
          description: Record stored
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/stats:
    get:
      operationId: getPracticeStats
      summary: Practice totals, per-exercise accuracy, and a daily trend
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Aggregated practice statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/daily:
    get:
      operationId: listDailyHistory
      summary: Per-day practice history for the calendar
      description: |
        One bucket per local day, oldest first, including days with no practice.
        Each day is judged against the goal that was in force then; days from
        before goals were recorded fall back to the current goal.
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 7
            maximum: 400
            default: 371
      responses:
        '200':
          description: Daily buckets and streaks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/collections:
    get:
      operationId: listCollections
      summary: The user's collection folders
      description: |
        Folders of levels. The default folder is created on first read with an
        empty name, which the client renders in the reader's language.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Collections with the level slugs in each
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCollection
      summary: Create a collection folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRequest'
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/collections/{id}:
    patch:
      operationId: renameCollection
      summary: Rename a collection folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRequest'
      responses:
        '204':
          description: Renamed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such folder, or the default folder (which has no name)
    delete:
      operationId: deleteCollection
      summary: Delete a collection folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such folder, or the default folder
  /me/collections/{id}/levels/{levelId}:
    put:
      operationId: addCollectionItem
      summary: Put a level in a folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - name: levelId
          in: path
          required: true
          schema:
            type: string
            example: singleNote-1
      responses:
        '204':
          description: Added
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such folder for this user
    delete:
      operationId: removeCollectionItem
      summary: Take a level out of a folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - name: levelId
          in: path
          required: true
          schema:
            type: string
            example: singleNote-1
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such folder for this user
...
  /me/levels:
    get:
      operationId: listLevelProgress
      summary: Question-set progress for the current user
      description: |
        Levels themselves are product content defined by the client; this returns
        what the user has attempted and passed.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: One entry per attempted level
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LevelProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/levels/{id}:
    post:
      operationId: recordLevelResult
      summary: Record a finished question-set level attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: singleNote-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelResultRequest'
      responses:
        '200':
          description: Stored progress for this level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelProgress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/mistakes:
    get:
      operationId: listMistakes
      summary: List the open mistakes in the notebook
      description: Questions answered wrong and not answered correctly again yet.
      security:
        - bearerAuth: []
      parameters:
        - name: exercise
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ExerciseKind'
      responses:
        '200':
          description: Open mistakes, most recently missed first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mistake'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/mistakes/{id}:
    delete:
      operationId: resolveMistake
      summary: Remove a mistake from the notebook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Mistake cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such open mistake for this user
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExerciseKind:
      type: string
      description: The five exercise modules. Melody and rhythm are scored as a whole.
      enum:
        - singleNote
        - interval
        - chord
        - melody
        - rhythm
    DailyProgress:
      type: object
      required:
        - date
        - solved
        - correct
      properties:
        date:
          type: string
          format: date
          description: Calendar date in the server's APP_TIMEZONE.
        solved:
          type: integer
        correct:
          type: integer
        byExercise:
          type: object
          description: Solved count per exercise kind, only for kinds practised today.
          additionalProperties:
            type: integer
    StudyPlan:
      type: object
      required:
        - dailyGoal
        - focusExercises
        - today
      properties:
        dailyGoal:
          type: integer
          minimum: 1
          maximum: 500
        focusExercises:
          type: array
          description: |
            The modules the daily session draws from. Empty means no narrowing down, so every
            module counts. A plan that was never saved answers with the defaults.
          items:
            $ref: '#/components/schemas/ExerciseKind'
        today:
          $ref: '#/components/schemas/DailyProgress'
    StudyPlanUpdate:
      type: object
      required:
        - dailyGoal
      properties:
        dailyGoal:
          type: integer
          minimum: 1
          maximum: 500
        focusExercises:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseKind'
    PracticeRecordRequest:
      type: object
      required:
        - exercise
        - correct
      properties:
        exercise:
          $ref: '#/components/schemas/ExerciseKind'
        correct:
          type: boolean
        chosen:
          type: string
          nullable: true
        expected:
          type: string
          nullable: true
        prompt:
          type: object
          additionalProperties: true
          nullable: true
          description: |
            The question itself (notes, interval, chord type, rhythm pattern...).
            Wrong answers carrying a prompt land in the mistake notebook, and a
            later correct answer for the same prompt clears the entry.
    DailyBucket:
      type: object
      required:
        - date
        - solved
        - correct
        - goal
        - met
      properties:
        date:
          type: string
          format: date
        solved:
          type: integer
        correct:
          type: integer
        goal:
          type: integer
          description: The daily goal that applied that day.
        met:
          type: boolean
          description: Solved reached the goal.
    DailyHistory:
      type: object
      required:
        - days
        - currentStreak
        - longestStreak
      properties:
        days:
          type: array
          items:
            $ref: '#/components/schemas/DailyBucket'
        currentStreak:
          type: integer
          description: Consecutive days up to today (or yesterday) with any practice.
        longestStreak:
          type: integer
    Collection:
      type: object
      required:
        - id
        - name
        - isDefault
        - levels
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Empty for the default folder, which the client names.
        isDefault:
          type: boolean
        levels:
          type: array
          items:
            type: string
          description: Level slugs in this folder.
    CollectionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
...
    LevelProgress:
      type: object
      required:
        - levelId
        - module
        - passed
        - bestAccuracy
      properties:
        levelId:
          type: string
          example: singleNote-1
        module:
          $ref: '#/components/schemas/ExerciseKind'
        passed:
          type: boolean
        bestAccuracy:
          type: number
          format: double
          description: Best fraction of correct answers across attempts, 0–1.
    LevelResultRequest:
      type: object
      required:
        - module
        - correct
        - total
        - passMark
      properties:
        module:
          $ref: '#/components/schemas/ExerciseKind'
        correct:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        passMark:
          type: number
          format: double
          description: Accuracy needed to pass, 0–1. The server decides, not the client.
    Mistake:
      type: object
      required:
        - id
        - exercise
        - answer
        - wrongCount
        - lastWrongAt
      properties:
        id:
          type: string
          format: uuid
        exercise:
          $ref: '#/components/schemas/ExerciseKind'
        prompt:
          type: object
          additionalProperties: true
          nullable: true
        answer:
          type: string
        wrongCount:
          type: integer
        lastWrongAt:
          type: string
          format: date-time
    ExerciseStats:
      type: object
      required:
        - solved
        - correct
        - accuracy
      properties:
        solved:
          type: integer
        correct:
          type: integer
        accuracy:
          type: number
          format: double
          description: Fraction between 0 and 1; 0 when nothing was solved.
    PracticeStats:
      type: object
      required:
        - solved
        - correct
        - accuracy
        - byExercise
        - daily
        - streak
      properties:
        solved:
          type: integer
        correct:
          type: integer
        accuracy:
          type: number
          format: double
        streak:
          type: integer
          description: |
            Consecutive days ending today (or yesterday, when today has no
            practice yet) with at least one answered question.
        byExercise:
          type: object
          description: Keyed by exercise kind; kinds never practised are absent.
          additionalProperties:
            $ref: '#/components/schemas/ExerciseStats'
        daily:
          type: array
          description: One entry per calendar day, oldest first, gaps filled with zeros.
          items:
            $ref: '#/components/schemas/DailyProgress'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'